# BildixApi/bin/Debug/net9.0/openapi.yaml
...
openapi: 3.0.3
...
info:
  title: Bildix API
  version: 1.0.0
  description: |
    Recruitment platform API.
    Matches candidates with job offers using rule-based and semantic scoring.

# ── BLOC 2 : SERVEURS ─────────────────────────────────────────────────────────
# Où tourne l'API ? On peut en lister plusieurs (dev, staging, prod).
# Swagger UI propose une liste déroulante pour choisir contre quel serveur tester.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
  # - url: https://api.bildix.com/v1
  #   description: Production  (uncomment when deployed)

# ── BLOC 3 : SÉCURITÉ GLOBALE ─────────────────────────────────────────────────
# Déclare que TOUS les endpoints nécessitent un token JWT par défaut.
# Un endpoint peut surcharger cette règle avec "security: []" pour être public.
security:
  - bearerAuth: []

# ══════════════════════════════════════════════════════════════════════════════
# BLOC 4 : COMPOSANTS RÉUTILISABLES
# ══════════════════════════════════════════════════════════════════════════════
# Tout ce qui est défini ici peut être référencé ailleurs avec $ref.
# Exemple : $ref: '#/components/schemas/Candidate'
# Cela évite la répétition : on décrit une fois, on réutilise partout.
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Réponses d'erreur réutilisables ───────────────────────────────────────
  # Plutôt que répéter "401 Unauthorized" dans chaque endpoint, on le déclare
  # une fois ici et on y fait référence avec $ref.
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schémas de données ────────────────────────────────────────────────────
  # Un schéma décrit la FORME d'un objet JSON :
  # - quelles propriétés il a
  # - quel type chacune (string, integer, boolean, array, object…)
  # - lesquelles sont obligatoires (listed under "required")
  # - contraintes optionnelles (minLength, minimum, enum…)
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Candidate not found

    # ══════════════════════════════════════════════════════════════════════
    # CANDIDATE
    # ══════════════════════════════════════════════════════════════════════

    # Ce que l'API RETOURNE quand on lit un candidat.
    # Contient les champs calculés par le serveur (id, createdAt).
...
    Company:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Acme Corp
        sector:
          type: string
          example: Technology
        location:
          type: string
          example: Lyon, France
        websiteUrl:
          type: string
          format: uri
          example: https://acme.com
        createdAt:
          type: string
          format: date-time
    CreateCompanyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Acme Corp
        sector:
          type: string
          example: Technology
        location:
          type: string
          example: Lyon, France
        websiteUrl:
          type: string
          format: uri

    # ══════════════════════════════════════════════════════════════════════
    # JOB OFFER
    # ══════════════════════════════════════════════════════════════════════
...
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          minLength: 8
          example: "SecurePass1!"
        role:
          type: string
          enum: [candidate, company]
          example: candidate
        turnstileToken:
          type: string
          description: |
            Cloudflare Turnstile challenge token obtained from the frontend widget.
            Omit or leave empty when calling from Swagger UI (verification is skipped
            server-side when Turnstile:SecretKey is not configured).
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    # The token returned on successful login or registration.
    # The client stores this and sends it as: Authorization: Bearer <token>
    AuthResponse:
      type: object
      required: [token, expiresAt, userId, role]
      properties:
        token:
          type: string
          description: JWT bearer token — include in Authorization header for all protected requests
          example: eyJhbGciOiJIUzI1NiJ9...
        expiresAt:
          type: string
          format: date-time
          description: UTC timestamp when the token expires
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [candidate, company, admin]

# ══════════════════════════════════════════════════════════════════════════════
# BLOC 5 : PATHS (les endpoints)
# ══════════════════════════════════════════════════════════════════════════════
# Chaque "path" est une route HTTP. On y définit :
#   - le verbe HTTP (get, post, put, delete, patch)
#   - les paramètres (path params, query params)
#   - le corps de la requête (requestBody)
#   - les réponses possibles avec leur code HTTP et leur schéma
#
# Convention : les routes ne contiennent jamais de verbe.
#   ✅ GET /candidates      ❌ GET /getCandidates
#   ✅ POST /candidates     ❌ POST /createCandidate
paths:
...
  /companies:
    get:
      summary: List all companies
      tags: [Companies]
      responses:
        '200':
          description: Array of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create a company
      tags: [Companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Company created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /companies/{id}:
    get:
      summary: Get a company by ID
      tags: [Companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Company found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── JOB OFFERS ─────────────────────────────────────────────────────────────
...
  /auth/register:
    post:
      summary: Create a user account
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created — token returned immediately
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email address already in use
  /auth/login:
    post:
      summary: Log in and receive a JWT token
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid email or password